Add Fantasque Sans Mono No Loop K
The default font has a looped small case k. Some people don't mind this.
But some other people, like me, prefer the no loop version, which the
creator of the font provides as well. So this commit adds this option to
this Atom package so I don't have to manually patch the normal version
each time I install Atom in a new machine.